neki za apije

+++ Kino/Kino/Models/Naslov.cs
@@ -6,7 +6,7 @@ namespace Kino.Models
     public partial class Naslov
     {
         public decimal IdNaslov { get; set; }
-        public int StPoste { get; set; }
+        public decimal StPoste { get; set; }
         public string Ulica { get; set; }
         public decimal HisnaSt { get; set; }
     }
